+++ CHANGES.md
@@ -44,6 +44,7 @@ Released in June 2022. For more information about the previous status, go to pre
 * The way flow is coded has been improbed. Dataflows are not exposed anymore. Callbacks registered have explicitly the input arguments they require and when doing the registration, these are compare against the inputs configured in the Policy (being type safe then). This makes the code easier to understand.
 * Inputs can be configured as consumable or not.
 * The Plotter has been improved a lot! Now the data displayed is smoother and fits better to real data.
+* Added module for audio (microphone, speaker, DFT, etc.).
 
 
 ### Known Bugs